[RJAJ-6] Tests for `AbstractModelBuilder`
This covers 100% of the class just by testing the `public` methods. If we find gaps later, we can add extra tests for the `protected` methods, but I'm going to run with this for now.

Also adds Mockito, PowerMock, and PowerMockito, for spying on classes and mocking static methods. Unfortunately, we can't use the latest version of Mockito with PowerMock because it's not yet compatible with it.