[RJAJ-12] Disable "New Object in Loop" PMD Check
I understand that it can sometimes be a bad idea to instantiate new objects in loops -- especially if there are a lot of iterations of the loop and/or instantiating the objects in question is expensive. However, this should be left up to the developer's best judgment on a case-by-case basis because most of the time this should be find. In the cases it's not, let's let profiling tools rather than premature optimization tell us where we've gone wrong.