[RJAJ-12] Disable `AbstractNaming` PMD Inspection
There is no need to force repetition of `Abstract` in the names of abtract classes. The only case in which we should is a case in which the abstract class provides convenient base functionality for an interface, but that should be the only case.

For more reasons why we should not force this inspection, see:
https://softwareengineering.stackexchange.com/questions/152759/should-i-add-an-abstract-prefix-to-my-abstract-classes