[RJAJ-12] Disable Abstract Class PMD Inspection
I can find no literature outside of PMD that would explain why declaring a class abstract without abstract methods is a code smell. Furthermore, why is it preferable to make the constructor of such a class protected rather than making the class abstract, especially if the class does indeed have concrete sub-classes?

In addition, a class with a protected constructor is still instantiable via reflection while an abstract class is not. Marking a class abstract very clearly communicates that one does not expect to see an object of solely that type out in the wild -- you expect one of its sub-classes. But, if there exists the possibility of instantiating the object -- even if it requires reflection -- that invariant might not hold for the rest of the system.