[RJAJ-12] Suppress AccessorMethodGeneration in PMD
Accessing private constructors through builders is a design element in our projects, so we don't want to make the pattern more complicated just because we're ending up with synthetic accessor methods.

We can re-evaluate this if we need to slim down the framework to run on embedded or mobile platforms like Android that have method limits; or, if we discover a security concern with the code having synthetic methods generated with package-private access.