[RJAJ-12] Post merge: Correct `clone()` Warnings
The implementation of `clone()` offered by AbstractModel is not a complete implementation -- like the one provided by the `Object` class, we expect it to be overridden and made public. Therefore, the normal PMD inspections related to it are not applicable.

With respect to the clone method in the test model: there's no reason it should not follow the expected contract. So, now it does.