[RJAJ-12] Fix-up PMD Exception-related Violations
Fixes instances of AvoidCatchingGenericException and EmptyCatchBlock in the `AbstractModelTest`. In general, these inspections aren't useful in tests because we often are testing system behavior during and after an exception, so we don't need to care about handling the exception. However, in this instance, there were a few exceptions I needed to prune back.