[RJAJ-12] Qualify Field References with `this.`
Although this may seem superfluous, the goal is make all references to fields (including in constructors) consistent, so there's no question about where variables are defined. This also allows us to mostly skip the rule that we should not have a local var shadow a class var, since class vars must be qualified with `this.`.