[RJAJ-6] Eliminate Duplication in SCB using JsonUtils
JsonUtils used to be for tests only; but since the functionality was so useful to application code, it was promoted a while back. Since we have it as a tool at our disposal, it makes sense to use it so we have one fewer spot to maintain with serializer configuration.