[RJAJ-6] Fix-up CloningPreprocessor + Add Test
The expectation about whether or not a field pre-processor can receive `null` was previously not documented. We didn't encounter an issue with it, since ABMB instances don't invoke pre-processors when the field value is `null` (to avoid overhead), but there's technically no reason that pre-processors shouldn't be able to gracefully handle `null`.